+++ inc/class-bridge-library-resources.php
@@ -125,6 +125,9 @@ class Bridge_Library_Resources extends Bridge_Library {
 		// Cache data from Primo PNX API.
 		add_action( 'acf/save_post', array( $this, 'cache_metadata' ), 10 );
 
+		// Schedule automatic updates.
+		add_action( 'bridge_library_schedule_daily', array( $this, 'background_update_resources()' ), 10 );
+
 		// Add to department related resources.
 		add_action( 'acf/update_value/key=field_5cc327908d3f4', array( $this, 'save_department_resources' ), 8, 3 );
 
@@ -298,6 +301,18 @@ class Bridge_Library_Resources extends Bridge_Library {
 		$update = $wpdb->update( $wpdb->prefix . $this->acf_meta_table, array( 'related_courses_resources' => $value ), array( 'post_id' => $post_id ) ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery
 
 		return false !== $update;
+	}
+
+	/**
+	 * Run daily imports.
+	 *
+	 * @since 1.6.0
+	 *
+	 * @return void
+	 */
+	public function background_update_resources() {
+		$this->background_create_resource_from_libguides_assets( true );
+		$this->background_create_resource_from_libguides_guides( true );
 	}
 
 	/**
@@ -583,7 +598,7 @@ class Bridge_Library_Resources extends Bridge_Library {
 			wp_die();
 		}
 
-		$term_id = sanitize_key( $_REQUEST['data']['term_id'] );
+		$term_id = absint( $_REQUEST['data']['term_id'] );
 		$term    = get_term_by( 'id', $term_id, 'academic_department' );
 
 		$primo_api          = Bridge_Library_API_Primo::get_instance();
@@ -687,6 +702,8 @@ class Bridge_Library_Resources extends Bridge_Library {
 			$institution_term = get_term_by( 'slug', 'st-olaf', 'institution' );
 		} elseif ( 'carleton.edu' === $domain_name ) {
 			$institution_term = get_term_by( 'slug', 'carleton', 'institution' );
+		} else {
+			throw new RuntimeException( 'Invalid institution' );
 		}
 
 		$tax_input = array(
@@ -1074,6 +1091,8 @@ class Bridge_Library_Resources extends Bridge_Library {
 			$asset_id = null;
 		}
 
+		$updated = 0;
+
 		foreach ( $stolaf_results as $result ) {
 			if ( $asset_id === $result['id'] ) {
 				$updated = $this->create_resource_from_libguides_asset( $result, 'st-olaf' );
@@ -1134,6 +1153,9 @@ class Bridge_Library_Resources extends Bridge_Library {
 		$stolaf_results   = $libguides_api_12->set_institution( 'stolaf' )->get_assets();
 		$carleton_results = $libguides_api_12->set_institution( 'carleton' )->get_assets();
 
+		error_log( 'Libguides background update: retrieved ' . count( $stolaf_results ) . ' St. Olaf assets' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+		error_log( 'Libguides background update: retrieved ' . count( $carleton_results ) . ' Carleton assets' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+
 		// LibGuides doesn’t have a limit parameter, so we have to get everything in one call.
 		// If async, we’ll break it up into chunks and schedule them to be processed.
 		if ( $async ) {
@@ -1158,7 +1180,8 @@ class Bridge_Library_Resources extends Bridge_Library {
 			}
 
 			// Now start the queue.
-			return $libguides_api_12->async->save()->dispatch();
+			$libguides_api_12->async->save()->dispatch();
+			return;
 		} else {
 			$assets = array();
 			foreach ( $stolaf_results as $asset ) {
@@ -1214,6 +1237,7 @@ class Bridge_Library_Resources extends Bridge_Library {
 			$guide_id = absint( sanitize_key( $_REQUEST['libguides_guide_id'] ) );
 		} else {
 			wp_send_json_error( array( 'error' => __( 'You must specify a guide ID', 'bridge-library' ) ), 400 );
+			return;
 		}
 
 		$libguides_api_11 = Bridge_Library_API_LibGuides_11::get_instance();
@@ -1221,6 +1245,8 @@ class Bridge_Library_Resources extends Bridge_Library {
 		$stolaf_guide = $libguides_api_11->set_institution( 'stolaf' )->get_guide( $guide_id );
 
 		$carleton_guide = $libguides_api_11->set_institution( 'carleton' )->get_guide( $guide_id );
+
+		$updated = 0;
 
 		if ( $stolaf_guide ) {
 			$updated = $this->create_resource_from_libguides_guide( $stolaf_guide, 'st-olaf' );
@@ -1258,7 +1284,7 @@ class Bridge_Library_Resources extends Bridge_Library {
 
 		$post_id = absint( $_REQUEST['post_id'] );
 
-		$assets = $this->create_resources_for_libguides_guide( absint( $_REQUEST['libguides_guide_id'] ), false );
+		$assets = $this->create_resources_for_libguides_guide( absint( $_REQUEST['libguides_guide_id'] ) );
 
 		$core_resources = get_field( 'core_resources', $post_id );
 		if ( empty( $core_resources ) ) {
@@ -1286,6 +1312,9 @@ class Bridge_Library_Resources extends Bridge_Library {
 		$stolaf_results   = $libguides_api_11->set_institution( 'stolaf' )->get_guides();
 		$carleton_results = $libguides_api_11->set_institution( 'carleton' )->get_guides();
 
+		error_log( 'Libguides background update: retrieved ' . count( $stolaf_results ) . ' St. Olaf guides' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+		error_log( 'Libguides background update: retrieved ' . count( $carleton_results ) . ' Carleton guides' ); // phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
+
 		// LibGuides doesn’t have a limit parameter, so we have to get everything in one call.
 		// If async, we’ll break it up into chunks and schedule them to be processed.
 		if ( $async ) {
@@ -1310,7 +1339,8 @@ class Bridge_Library_Resources extends Bridge_Library {
 			}
 
 			// Now start the queue.
-			return $libguides_api_11->async->save()->dispatch();
+			$libguides_api_11->async->save()->dispatch();
+			return;
 		} else {
 			$results = array();
 			foreach ( $stolaf_results as $guide ) {