+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] — 2023-09-20
+
+- Feature: add daily LibGuides assets import
+- Feature: add daily LibGuides guides import
+- Feature: add LibGuides error monitoring
+
 ## [1.5.0] — 2023-09-20
 
 - Feature: add ability to retrieve courses by specific course terms